tasks crud guard + relationship

+++ workspaces/api/src/tasks/tasks.service.ts
@@ -1,6 +1,8 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskStatus } from '@tasks-manager/common';
+import { User } from 'src/auth/user.entity';
+
 import { TaskDTO, TaskFilterDTO } from './task.dto';
 import { Task } from './task.entity';
 import { TaskRepository } from './task.repository';
@@ -12,12 +14,14 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDTO: TaskFilterDTO): Promise<Task[]> {
-    return await this.taskRepository.getTasks(filterDTO);
+  async getTasks(filterDTO: TaskFilterDTO, user: User): Promise<Task[]> {
+    return await this.taskRepository.getTasks(filterDTO, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, ownerId: user.id },
+    });
 
     if (!found) {
       throw new NotFoundException(`Task with id ${id} not found`);
@@ -26,19 +30,25 @@ export class TasksService {
     return found;
   }
 
-  async createTask(taskDTO: TaskDTO): Promise<Task> {
+  async createTask(taskDTO: TaskDTO, user: User): Promise<Task> {
     const task = await this.taskRepository.create({
       ...taskDTO,
       status: TaskStatus.OPEN,
+      owner: user,
     });
 
     await this.taskRepository.save(task);
 
+    delete task.owner;
     return task;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
 
     await task.save();
@@ -46,8 +56,8 @@ export class TasksService {
     return task;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const res = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const res = await this.taskRepository.delete({ id, ownerId: user.id });
 
     if (!res.affected) {
       throw new NotFoundException(`Task with id ${id} not found`);
